+++ sktime/utils/time_series.py
@@ -1,11 +1,20 @@
+__author__ = ["Markus Löning"]
+__all__ = [
+    "compute_n_intervals",
+    "time_series_slope",
+    "fit_trend",
+    "remove_trend",
+    "add_trend",
+    "split_into_tabular_train_test"
+]
+
 import numpy as np
 from sklearn.utils import check_array
 
-from sktime.forecasters.model_selection import RollingWindowSplit
 from sktime.utils.validation.forecasting import validate_fh, validate_time_index
 
 
-def get_n_intervals(n_timepoints, n_intervals="sqrt"):
+def compute_n_intervals(n_timepoints, n_intervals="sqrt"):
     """
     Compute number of intervals for given number of time points
 
@@ -240,41 +249,6 @@ def add_trend(x, coefs, time_index=None):
 
     return xt
 
-
-def split_into_tabular_train_test(x, window_length=None, fh=None, test_size=1):
-    """Helper function to split single time series into tabular train and
-    test sets using rolling window approach"""
-
-    # validate forecasting horizon
-    fh = validate_fh(fh)
-
-    # get time series index
-    index = np.arange(len(x))
-
-    # set up rolling window iterator
-    rw = RollingWindowSplit(window_length=window_length, fh=fh)
-
-    # slice time series into windows
-    xs = []
-    ys = []
-    for input, output in rw.split(index):
-        xt = x[input]
-        yt = x[output]
-        xs.append(xt)
-        ys.append(yt)
-
-    # stack windows into tabular array
-    x = np.array(xs)
-    y = np.array(ys)
-
-    # split into train and test set
-    x_train = x[:-test_size, :]
-    y_train = y[:-test_size, :]
-
-    x_test = x[-test_size:, :]
-    y_test = y[-test_size:, :]
-
-    return x_train, y_train, x_test, y_test
